+++ documentation/.~lock.setup.odt#
@@ -1,0 +1,1 @@
+,gmiak,gmiak-HP-ProBook-430-G2,03.03.2022 21:39,file:///home/gmiak/.config/libreoffice/4;
