+++ server/package.json
@@ -20,7 +20,9 @@
   "homepage": "https://github.com/gmiak/movieLib#readme",
   "dependencies": {
     "cors": "^2.8.5",
-    "express": "^4.17.2"
+    "express": "^4.17.2",
+    "mongodb": "^4.4.1",
+    "mongoose": "^6.2.4"
   },
   "devDependencies": {
     "@babel/core": "^7.17.2",
@@ -28,6 +30,8 @@
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/jest": "^27.4.0",
+    "@types/mongodb": "^4.0.7",
+    "@types/mongoose": "^5.11.97",
     "@types/node": "^17.0.16",
     "@types/supertest": "^2.0.11",
     "babel-jest": "^27.5.1",
